Rate limiting refactor - moved into separate middleware file
