+++ src/llm-service/index.js
@@ -3,6 +3,7 @@ import axios from "axios";
 import cors from "cors";
 import multer from "multer";
 import dotenv from "dotenv";
+import { rateLimiter } from "./middleware/rateLimiter.js";
 
 // Load environment variables
 dotenv.config();
@@ -68,37 +69,6 @@ const extractFileContent = (file) => {
   } catch (error) {
     throw new Error('Failed to extract file content: ' + error.message);
   }
-};
-
-// Rate limiting parameters
-const rateLimit = {};
-const RATE_LIMIT = 10;
-const TIME_WINDOW = 60 * 1000;
-
-// Custom rate liminiting middleware
-const rateLimiter = (req, res, next) => {
-  const clientIP = req.ip || req.connection.remoteAddress;
-  const now = Date.now();
-  
-  if (!rateLimit[clientIP]) {
-    rateLimit[clientIP] = { count: 1, resetTime: now + TIME_WINDOW };
-    return next();
-  }
-  
-  if (now > rateLimit[clientIP].resetTime) {
-    rateLimit[clientIP] = { count: 1, resetTime: now + TIME_WINDOW };
-    return next();
-  }
-  
-  if (rateLimit[clientIP].count >= RATE_LIMIT) {
-    return res.status(429).json({ 
-      error: "Too many requests. Try again later.",
-      retryAfter: Math.ceil((rateLimit[clientIP].resetTime - now) / 1000)
-    });
-  }
-  
-  rateLimit[clientIP].count++;
-  next();
 };
 
 // Make sure to add API key in your .env file
